cleaned up

+++ src/lib/year_2025/day_01/get_input.rs
@@ -1,5 +1,5 @@
 #[allow(dead_code)]
-pub fn get_input() -> Vec<i32> {
+pub fn get_input() -> Vec<isize> {
     include_str!("input.txt")
         .lines()
         .map(|x| x.replace("R", "").replace("L", "-").parse().unwrap())
